requirements.txt
# Core
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1

# Database
psycopg2-binary==2.9.9
sqlalchemy>=1.4.28,<2.0  # Compatible con Airflow 2.7.3
alembic==1.13.0
redis==5.0.1

# Data processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
pdfplumber==0.10.3

# ML/Analytics
scikit-learn==1.3.2
statsmodels==0.14.0
prophet==1.1.5

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1

# Airflow
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1

# Visualization/Dashboard
streamlit==1.28.2
plotly==5.18.0
altair==5.1.2

# Utils
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
faker==20.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
ipython==8.18.0
jupyter==1.0.0
